Adding a tracker pre-seed file to fix wrong heuristic classifications and bootstrap tracker detection. Fixes #965
