+++ data/preseed.json
@@ -1,0 +1,3 @@
+{
+  "vk.me": "cookieblock"
+}
