+++ src/background.js
@@ -47,6 +47,7 @@ function Badger() {
       badger.runMigrations();
     } finally {
       badger.initializeCookieBlockList();
+      badger.initializePreSeed();
       badger.initializeDNT();
       badger.initializeUserAllowList();
       badger.enableWebRTCProtection();
@@ -187,6 +188,14 @@ Badger.prototype = {
         function(){ // empty callback
         });
     });
+  },
+
+   /**
+   * Initialize preSeed data (pre-configured tracker settings). To prevent over-blocking and to be able
+   * to initialize the most common trackers
+   */
+  initializePreSeed: function(){
+    this.storage.preSeedData = utils.loadJSONFromFile("data/preseed.json");
   },
 
   /**
